Fix comment_deletebutton and Add comment author

+++ back/src/services/userService.js
@@ -181,7 +181,27 @@ class userAuthService {
 
     return updatedLike;
   }
-  static async deleteUser({ userId }) {
+  static async deleteUser({ userId, email, password }) {
+    const user = await User.findByEmail({ email });
+    if (!user) {
+      const errorMessage =
+        "해당 이메일은 가입 내역이 없습니다. 다시 한 번 확인해 주세요.";
+      return { errorMessage };
+    }
+
+    // 비밀번호 일치 여부 확인
+    const correctPasswordHash = user.password;
+    const isPasswordCorrect = await bcrypt.compare(
+      password,
+      correctPasswordHash
+    );
+
+    if (!isPasswordCorrect) {
+      const errorMessage =
+        "비밀번호가 일치하지 않습니다. 다시 한 번 확인해 주세요.";
+      return { errorMessage };
+    }
+
     const isDataDeleted = await User.deleteById({ userId });
 
     // db에서 찾지 못한 경우, 에러 메시지 반환
